add resources

+++ cs544-202403-project/src/main/java/edu/miu/cs/cs544/service/EventService.java
@@ -4,6 +4,6 @@ import edu.miu.common.service.BaseReadWriteService;
 import edu.miu.cs.cs544.domain.Event;
 import edu.miu.cs.cs544.service.contract.EventPayload;
 
-public interface EventService extends BaseReadWriteService<EventPayload, Event, Long> {
+public interface EventService extends BaseReadWriteService<EventPayload, Event, Integer> {
     Event getByEventId(Integer eventId);
 }
